Containerize the k8s CNI plugin.
Create a container to run the calico_kubernetes_cni plugin.

The container can be built with `make plugin`, which builds
a docker container called `calico/kubernetes-cni`.

For now this container is based on Debian, and still calls
the calicoctl binary. We should optimize this by removing all
shell-outs in the CNI plugin, and then rebasing the docker image
onto a minimalist Python baseimage.